# openapi/main.yml
openapi: 3.0.3
info:
  title: Feedback API
  description: |
    フィードバックシステムのAPI仕様書です。
    Chrome拡張機能からのフィードバック送信、スクリーンショットのアップロード、
    エラーログの記録などの機能を提供します。
  version: 1.0.0
  contact:
    name: API Support
    email: dev@porhy.com
servers:
  - url: https://feedback-suite.polyrhythm.tokyo
    description: Production server
...
paths:
  /api/feedback:
    post:
      tags:
        - feedback
      summary: フィードバックを送信
      description: Chrome拡張機能からフィードバックを受信し、必要に応じてGitHub Issueやタスクサーバーにもデータを送信します
      operationId: createFeedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreateRequest'
      responses:
        '200':
          description: フィードバックを正常に受信しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackCreateResponse'
        '400':
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags:
        - feedback
      summary: CORS プリフライトリクエスト
      responses:
        '200':
          description: OK
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
                example: '*'
            Access-Control-Allow-Methods:
              schema:
                type: string
                example: 'GET, POST, PUT, DELETE, OPTIONS'
            Access-Control-Allow-Headers:
              schema:
                type: string
                example: 'Content-Type, Authorization'
...
  /api/feedback/{id}/create-task:
    post:
      tags:
        - feedback
      summary: フィードバックからタスクを作成
      description: 指定されたフィードバックIDからタスク管理サーバーにタスクを作成します
      operationId: createTaskFromFeedback
      parameters:
        - name: id
          in: path
          required: true
          description: フィードバックID
          schema:
            type: integer
      responses:
        '200':
          description: タスクが作成されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCreateResponse'
        '400':
          description: 無効なフィードバックID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: フィードバックまたはスクリーンショットデータが見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: タスク管理サーバーのAPIキーが設定されていません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags:
        - feedback
      summary: CORS プリフライトリクエスト
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
...
components:
  schemas:
    FeedbackCreateRequest:
      type: object
      required:
        - comment
      properties:
        comment:
          type: string
          description: フィードバックコメント
        uploadedDataId:
          type: string
          description: アップロード済みのスクリーンショットデータID
        timestamp:
          type: integer
          format: int64
          description: タイムスタンプ（ミリ秒）
        url:
          type: string
          description: フィードバックが送信されたページのURL
        githubRepository:
          type: string
          description: 関連するGitHubリポジトリ（例：owner/repo）
        errorDetails:
          type: object
          description: エラーの詳細情報。pageUrlを含む任意のキー・値を受け付けます
          additionalProperties: true
          properties:
            pageUrl:
              type: string
              description: エラーが発生したページのURL
    FeedbackCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        id:
          type: integer
          description: 作成されたフィードバックのID
        message:
          type: string
          example: フィードバックを受信しました
...
    TaskCreateResponse:
      type: object
      properties:
        success:
          type: boolean
        taskId:
          type: string
          description: 作成されたタスクのID
        taskUrl:
          type: string
          description: タスクの詳細URL
        message:
          type: string
          example: タスクが作成されました
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        details:
          type: string
          description: エラーの詳細情報
        success:
          type: boolean
          example: false